Fixed the bug where Special Auth tab was visible with no special authorizations.
